requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
httpx>=0.23.0
pydantic==2.6.4
pydantic-settings==2.6.1
supabase>=0.0.2
psycopg2-binary>=2.9.6
redis>=4.3.4
python-dotenv>=0.19.0
asyncio>=3.4.3
aiohttp>=3.8.1
pytest>=6.2.5
pytest-asyncio>=0.16.0
black>=21.9b0
isort>=5.9.3
mypy>=0.910
sentry-sdk>=1.5.0
APScheduler>=3.9.1
matplotlib>=3.5.0
pandas>=1.5.0
python-telegram-bot>=20.0.0,<21.0.0
